requirements-dev.txt
pytest
pytest-asyncio
freezegun
discord.py>=2,<3
